requirements.txt
anthropic
python-docx
python-dotenv
openai
pandas
playwright
PyPDF2
requests
streamlit
textract